[giac] Attempt to remove var=var in solve output. Fixed missing - in a test for regroup to work properly.
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@25927 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/giac/global.cc
@@ -3857,7 +3857,7 @@ unsigned int ConvertUTF16toUTF8 (
     }
 
     target += bytesToWrite;
-    if (target > targetEnd) {
+    if ((unsigned long)target > (unsigned long)targetEnd) {
         source = oldSource; /* Back up source pointer! */
         target -= bytesToWrite; result = targetExhausted; break;
     }
@@ -3958,7 +3958,7 @@ unsigned int ConvertUTF8toUTF16 (
     }
     ch -= offsetsFromUTF8[extraBytesToRead];
 
-    if (target >= targetEnd) {
+    if ((unsigned long)target >= (unsigned long)targetEnd) {
         source -= (extraBytesToRead+1); /* Back up source pointer! */
         result = targetExhausted; break;
     }
@@ -3989,7 +3989,7 @@ unsigned int ConvertUTF8toUTF16 (
         }
     } else {
         /* target is a character in range 0xFFFF - 0x10FFFF. */
-        if (target + 1 >= targetEnd) {
+        if ((unsigned long)target + 1 >= (unsigned long)targetEnd) {
         source -= (extraBytesToRead+1); /* Back up source pointer! */
         result = targetExhausted; break;
         }
